moved plot_comparison to its own file so it is separate from existing functions
